Refactor examples into separate translation units
