+++ src/examples/CMakeLists.txt
@@ -3,7 +3,15 @@
 # This file is part of the action_graph library and is licensed under the MIT
 # License. See the LICENSE file in the root directory for full license text.
 
-add_executable(action_graph_examples examples.cpp)
+add_executable(
+  action_graph_examples
+  examples_main.cpp
+  console_log.cpp
+  logging_builder.cpp
+  single_second_example.cpp
+  high_frequency_example.cpp
+  composite_graph_example.cpp
+  monitored_graph_example.cpp)
 
 target_link_libraries(
   action_graph_examples
