+++ src/examples/examples.h
@@ -1,0 +1,15 @@
+#ifndef ACTION_GRAPH_EXAMPLES_EXAMPLES_H
+#define ACTION_GRAPH_EXAMPLES_EXAMPLES_H
+
+#include "console_log.h"
+
+namespace action_graph_examples {
+
+void RunSingleSecondTriggerExample(ConsoleLog &log);
+void RunHighFrequencyTriggerExample(ConsoleLog &log);
+void RunCompositeGraphOnce(ConsoleLog &log);
+void RunMonitoredGraph(ConsoleLog &log);
+
+} // namespace action_graph_examples
+
+#endif // ACTION_GRAPH_EXAMPLES_EXAMPLES_H
